requirements.txt
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
filelock==3.13.1
idna==3.6
packaging==23.2
platformdirs==3.11.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
snowflake-connector-python==3.7.1
sortedcontainers==2.4.0
tomlkit==0.12.4
typing_extensions==4.10.0
urllib3==2.2.1
